a2a_utilities/requirements.txt
google-adk[a2a]>=1.8.0
python-dotenv
